third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-scroll-snap/snap-after-relayout/multiple-aligned-targets/prefer-focused-nested-containers-expected.txt

This is a testharness.js-based test.
[FAIL] Snap container prefers focused nested snap target.
  assert_equals: vertical scroll done expected 500 but got 494
Harness: the test ran to completion.
